+++ app/gui/src/project-view/components/GraphEditor/widgets/WidgetFunction.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { useCurrentProject, useProjectStore } from '$/components/WithCurrentProject.vue'
+import { useCurrentProject } from '$/components/WithCurrentProject.vue'
 import type { MethodCallInfo } from '$/providers/openedProjects/graph/graphDatabase'
 import {
   Score,
@@ -10,7 +10,10 @@ import {
   type WidgetUpdate,
 } from '$/providers/openedProjects/widgetRegistry'
 import NodeWidget from '@/components/GraphEditor/NodeWidget.vue'
-import { useWidgetFunctionCallInfo } from '@/components/GraphEditor/widgets/WidgetFunction/widgetFunctionCallInfo'
+import {
+  getPotentialModuleFunctionPointer,
+  useWidgetFunctionCallInfo,
+} from '@/components/GraphEditor/widgets/WidgetFunction/widgetFunctionCallInfo'
 import { injectFunctionInfo, provideFunctionInfo } from '@/providers/functionInfo'
 import { assert, assertUnreachable } from '@/util/assert'
 import { Ast } from '@/util/ast'
@@ -28,10 +31,16 @@ import { isIdentifier } from '@/util/qualifiedName'
 import { proxyRefs } from '@/util/reactivity'
 import { computed } from 'vue'
 import { Err, Ok } from 'ydoc-shared/util/data/result'
+import { FunctionName } from './WidgetFunctionName.vue'
 
 const props = defineProps(widgetProps(widgetDefinition))
-const { projectNames: projectNames, module, graph } = useCurrentProject()
-const project = useProjectStore()
+const {
+  store: project,
+  projectNames: projectNames,
+  module,
+  graph,
+  suggestionDb,
+} = useCurrentProject()
 
 const exprInfo = computed(() => graph.value.db.getExpressionInfo(props.input.value.externalId))
 const outputType = computed(() => exprInfo.value?.typeInfo?.primaryType)
@@ -41,6 +50,7 @@ const { methodCallInfo, application, subject, subjectInfo } = useWidgetFunctionC
   () => graph.value.db,
   project,
   projectNames,
+  () => suggestionDb.value.groups,
 )
 
 provideFunctionInfo(
@@ -69,7 +79,23 @@ const innerInput = computed(() => {
     input = { ...props.input }
   }
   const callInfo = methodCallInfo.value
-  if (callInfo) input[CallInfo] = callInfo
+  if (callInfo) {
+    input[CallInfo] = callInfo
+    if (
+      !(ArgumentApplicationKey in input) &&
+      (input.value instanceof Ast.PropertyAccess || input.value instanceof Ast.Ident)
+    ) {
+      const methodPointer = callInfo.methodCall.methodPointer
+      const definition = module.value.getMethodAst(methodPointer)
+      if (definition.ok) {
+        input[FunctionName] = {
+          editableNameExpression: definition.value.name.externalId,
+          methodPointer,
+          requireUserAction: true,
+        }
+      }
+    }
+  }
   return input
 })
 
@@ -257,9 +283,29 @@ export const widgetDefinition = defineWidget(
       if (prevFunctionState?.prefixCalls.has(ast.id)) return Score.Mismatch
 
       if (ast instanceof Ast.App || ast instanceof Ast.OprApp) return Score.Perfect
+      if (getMethodCallInfoRecursively(ast, db)) return Score.Perfect
+      if (
+        ast instanceof Ast.PropertyAccess ||
+        (ast instanceof Ast.Ident && !(ast.parent() instanceof Ast.PropertyAccess))
+      ) {
+        // Apply WidgetFunction to simple non-application expressions that represent an existing module-local function,
+        // even if it wasn't evaluated yet (i.e. we don't have up-to-date expression data yet).
+        const currentProject = useCurrentProject(true)
+        const projectPath = currentProject?.store.value.moduleProjectPath
+        if (currentProject && projectPath?.ok) {
+          const names = currentProject.projectNames
+          const potentialMethod = getPotentialModuleFunctionPointer(
+            ast,
+            projectPath.value,
+            names,
+            db,
+          )
+          if (potentialMethod && currentProject.module.value.hasMethod(potentialMethod.name))
+            return Score.Perfect
+        }
+      }
 
-      const info = getMethodCallInfoRecursively(ast, db)
-      return info != null ? Score.Perfect : Score.Mismatch
+      return Score.Mismatch
     },
   },
   import.meta.hot,