+++ app/gui/src/project-view/providers/asyncResources/upload.ts
@@ -1,9 +1,15 @@
 import { type Asset, AssetType, DirectoryId } from '#/services/Backend'
 import RemoteBackend from '#/services/RemoteBackend'
 import { unsafeKeys } from '#/utilities/object'
+<<<<<<< HEAD
 import type { OpenedProject, OpenedProjectsStore } from '$/providers/openedProjects'
 import { getFolderPath, readUserSelectedFile } from '$/utils/file'
 import { backendMutationOptions, backendQueryOptions } from '@/composables/backend'
+=======
+import type { OpenedProjectsStore } from '$/providers/openedProjects'
+import type { Initialized as InitializedProject } from '$/providers/openedProjects/projectStates'
+import { readUserSelectedFile } from '$/utils/file'
+>>>>>>> 6f062b5923752c661acff933fbd19acd3e044ff9
 import { useProjectFiles } from '@/stores/projectFiles'
 import { Err, mapOk, Ok, type Result } from '@/util/data/result'
 import { QueryClient, useMutation } from '@tanstack/vue-query'
@@ -102,7 +108,7 @@ export function useResourceUpload(
   const uploadImageMutation = useMutation(backendMutationOptions('uploadImage', backend), query)
 
   async function uploadResourceToProject(
-    project: OpenedProject,
+    project: InitializedProject,
     upload: UploadDefinition,
   ): Promise<Result<UploadProgress>> {
     const api = useProjectFiles(project.store)
@@ -167,8 +173,13 @@ export function useResourceUpload(
     context: ResourceContextSnapshot,
   ): Promise<Result<UploadProgress>> {
     const openedProject = context.project && openedProjects.get(context.project)
-    if (openedProject) {
-      return uploadResourceToProject(openedProject, data)
+    if (openedProject?.nextTask?.process === 'opening') {
+      await openedProjects.waitForProcess(openedProject)
+    }
+    const initialized =
+      openedProject?.state.status === 'initialized' ? openedProject.state : undefined
+    if (initialized) {
+      return uploadResourceToProject(initialized, data)
     } else if (context.asset) {
       return uploadResourceToCloud(data, context.asset)
     } else {