+++ engine/runtime/src/main/scala/org/enso/interpreter/runtime/DefaultPackageRepository.scala
@@ -212,7 +212,7 @@ private class DefaultPackageRepository(
     isLibrary: Boolean
   ): Unit = {
     val extensions = pkg.listPolyglotExtensions("java")
-    extensions.foreach(context.addToClassPath(pkg, _))
+    extensions.foreach(context.addToClassPath(pkg, _, false))
 
     val (regularModules, syntheticModulesMetadata) = pkg
       .listSources()